+++ CHANGELOG.md
@@ -1,6 +1,33 @@
 cookbook-rb-manager CHANGELOG
 ===============
 
+## 5.16.0
+
+  - Pablo Pérez
+    - [3e29e14] Increment aerospike memory
+    - [565751c] add defatult malware loaders
+    - [20dbb9d] Added cron task create_aerospike_secondary_indexes_daily
+    - [7b93c26] Add policy enforced data into node
+
+## 5.15.3
+
+  - nilsver
+    - [b8e643d] Bugfix/20613 sflow locations show double if same ip (#287)
+
+## 5.15.2
+
+  - Rafael Gomez
+    - [34163b6] Change ipaddress variable to ipaddress_sync
+  - manegron
+    - [f462847] Fix aerospike
+
+## 5.15.1
+
+  - nilsver
+    - [3184297] fix from other task iptables
+  - ljblancoredborder
+    - [6281f9d] adding nginx as a vip and as balanced service for clustering
+
 ## 5.15.0
 
   - nilsver